Fix sending token value based on decimals
